--- a/src.cmake
+++ b/src.cmake
@@ -102,6 +102,10 @@ set(RENDERERLIST
     ${ENGINE_DIR}/renderer/tr_font.cpp
     ${ENGINE_DIR}/renderer/InternalImage.cpp
     ${ENGINE_DIR}/renderer/InternalImage.h
+    ${ENGINE_DIR}/renderer/Material.cpp
+    ${ENGINE_DIR}/renderer/Material.h
+    ${ENGINE_DIR}/renderer/TextureManager.cpp
+    ${ENGINE_DIR}/renderer/TextureManager.h
     ${ENGINE_DIR}/renderer/tr_image.cpp
     ${ENGINE_DIR}/renderer/tr_image.h
     ${ENGINE_DIR}/renderer/tr_image_crn.cpp
@@ -141,6 +145,8 @@ set(RENDERERLIST
 )
 
 set(GLSLSOURCELIST
+    ${ENGINE_DIR}/renderer/glsl_source/material_vp.glsl
+    ${ENGINE_DIR}/renderer/glsl_source/material_fp.glsl
     ${ENGINE_DIR}/renderer/glsl_source/skybox_vp.glsl
     ${ENGINE_DIR}/renderer/glsl_source/ssao_fp.glsl
     ${ENGINE_DIR}/renderer/glsl_source/ssao_vp.glsl