--- a/src/engine/renderer/glsl_source/heatHaze_fp.glsl
+++ b/src/engine/renderer/glsl_source/heatHaze_fp.glsl
@@ -22,6 +22,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 /* heatHaze_fp.glsl */
 
+#define HEATHAZE_GLSL
+
 uniform sampler2D	u_CurrentMap;
 uniform float		u_AlphaThreshold;
 
@@ -32,10 +34,16 @@ DECLARE_OUTPUT(vec4)
 
 void	main()
 {
+	#insert material_fp
+
 	vec4 color;
 
 	// compute normal in tangent space from normalmap
-	vec3 normal = NormalInTangentSpace(var_TexCoords);
+	#if defined(r_normalMapping)
+		vec3 normal = NormalInTangentSpace(var_TexCoords, u_NormalMap);
+	#else // !r_normalMapping
+		vec3 normal = NormalInTangentSpace(var_TexCoords);
+	#endif // !r_normalMapping
 
 	// calculate the screen texcoord in the 0.0 to 1.0 range
 	vec2 st = gl_FragCoord.st / r_FBufSize;