--- a/src/engine/renderer/glsl_source/liquid_vp.glsl
+++ b/src/engine/renderer/glsl_source/liquid_vp.glsl
@@ -40,6 +40,8 @@ OUT(smooth) vec3	var_Normal;
 
 void	main()
 {
+	#insert material_vp
+
 	// transform vertex position into homogenous clip-space
 	gl_Position = u_ModelViewProjectionMatrix * vec4(attr_Position, 1.0);