… src/engine/renderer/Material.cpp
… src/engine/renderer/gl_shader.cpp
… src/engine/renderer/gl_shader.h
… src/engine/renderer/glsl_source/liquid_fp.glsl
… src/engine/renderer/glsl_source/material_fp.glsl
… src/engine/renderer/glsl_source/material_vp.glsl
… src/engine/renderer/glsl_source/reliefMapping_fp.glsl
… src/engine/renderer/glsl_source/screen_fp.glsl
… src/engine/renderer/glsl_source/screen_vp.glsl
… src/engine/renderer/glsl_source/skybox_fp.glsl
… src/engine/renderer/glsl_source/skybox_vp.glsl
… src/engine/renderer/glsl_source/vertexSprite_vp.glsl
… src/engine/renderer/shaders.cpp
… src/engine/renderer/tr_backend.cpp
… src/engine/renderer/tr_bsp.cpp
… src/engine/renderer/tr_image.cpp
… src/engine/renderer/tr_init.cpp
… src/engine/renderer/tr_local.h
… src/engine/renderer/tr_main.cpp
… src/engine/renderer/tr_public.h
… src/engine/renderer/tr_scene.cpp
… src/engine/renderer/tr_shade.cpp
… src/engine/renderer/tr_sky.cpp
… src/engine/renderer/tr_surface.cpp
… src/engine/renderer/tr_vbo.cpp
… src/engine/renderer/tr_world.cpp
… src/engine/sys/sdl_glimp.cpp