Base changes for implementing a material system.

The goal of this system is to improve performance by reducing the amount of state changes and data uploads from CPU->GPU.

The system works by generating materials for draw surfaces/shaders/shader stages after the map is loaded, as well as all the necessary data to draw each material in one drawcall. Currently this only works for world BSP surfaces, everything else is drawn the normal way.

## Material
A material in this case represents a distinct global OpenGL state: this includes state bits (as used by GL_State(), however without the alphaTest bits), culling mode, shader program, material ordering and VBO/IBO bindings. Each material is also allocated 2 blocks of memory in a buffer, this memory holds the data for each surface.

## Material system
The material system holds material packs, which map materials to various shader sorts: depth pre-pass, opaque pass, and translucent pass.
It also contains sky shaders and a list of dynamic surfaces.

A dynamic surface is a surface which can have its data changed after the map is loaded. This includes non-constant color, tex matrices, animated and video images, as well as any active expressions used by RB_EvalExpression().

All surface data is held in a shader storage buffer object. This buffer consists of 2 blocks which have the following memory layout:
Material0_surface0_stage0: uniform0 // Start of material 0 data
..
Material0_surface0_stage0: uniformn
Material0_surface0_stage0: optional padding
Material0_surface0_stage1: uniform0
..
Material0_surface0_stagex: uniformn
Material0_surface0_stagex: optional padding
..
Material1: optional padding
Material1_surface0_stage0: uniform0
etc.

The optional padding for per-surface data is needed for it to align with std430 layout. The optional padding for materials is needed because each material can have a different data size per surface,  but the offset in the shader has to be an integer multiple of its materials size. The shader accesses the buffer using the offset that equals to: `bufferOffset / paddedSize`, `bufferOffset` is the offset in components i. e. multiples of 4 bytes because that's the minimum accessible memory in std430, `paddedSize` is the size of this shaders Material struct in components, aligned under std430.

The buffer is split into a static surface block, followed by a dynamic surface block. The dynamic surface block is updated once per frame. The whole block has to be updated because the BSP is entirely skipped for world surfaces.

There are 4 distinct steps done by the material system:
1. Shader post-processing. This will move the uniforms around for shaders marked as using the material system, and then generate the memory layout data (offsets and sizes), as well as generate code that will be inserted into the shader code instead of the uniforms. Uniforms marked as global are skipped: these uniforms are set per-shader rather than per-surface because they don't change within a rendered view.
2. Material generation. This will iterate through all of the world BSP surfaces except sky and portal surfaces using R_AddWorldSurfaces(). Then for each stage material data is obtained and if a matching material doesn't exist it will be added, then the surface data is added to the material. Surfaces will also be marked as either static or dynamic. Portal surfaces are copied so they can be rendered later, but until GPU culling is there it has to be done in an inefficient way.
3. Buffer generation. This will generate the buffer with all of the per-surface data, and then the command buffer.
4. Rendering. The material rendering is called from RB_RenderDrawSurfaces() with appropriate shader sorts. All of the materials matching this sort will the nbe rendered in order.

The command buffer contains data for multiDrawIndirect drawcalls. This contains the indexes into vertex and index array buffers, as well as baseInstance, which is used to access the buffer block in the shader. This data is static.

When rendering, all of the surfaces of a rendered material will be rendered with one such call.

While this means that all of the map surfaces are rendered every frame, it ends being much faster than walking the BSP tree and incurring all the state changes for maps with worse vis. Some maps that are good for testing this are https://users.unvanquished.net/~sweet/pkg/map-station12-b7_0.dpk and https://users.unvanquished.net/~kai/map-example_0.1.dpk.
The impact may vary depending on the device, but because currently at least on some devices perf can be worse with this system enabled depending on map, it is disabled by default. I expect that to work better or on par with what we currently have once GPU culling is added.

It can also later be extended to entities etc.

Some example performance differences:
map|scene|FPS(material system on)|FPS(master branch)
-|-|-|-
nova|-1896 336 248 0 0|210|170
habitat|default spectator view|200|150
station12-b7|-1377 -195 10 6 8|110|70
example|22728 21405 25414 -139 45|170|43

There are a few issues that are out of scope of this pr: proper portal culling, heatHaze stages (the only map that I know of that use heatHaze is `usstremor`, but there the heatHaze stage is on a non-world entity, so it works fine anyway), quakeFog stage.

To use the material system, do `/r_arb_bindless_texture 1; r_useMaterialSystem 1; vid_restart`.