@illwieckz 2024-04-27
I get a lot of those warnings:

```c++
Unvanquished/daemon/src/engine/renderer/gl_shader.h:478:37: warning: '&&' within '||' [-Wlogical-op-parentheses]
                if ( _shader->UseMaterialSystem() && _global || !_shader->UseMaterialSystem() ) {
                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~ ~~
```