@slipher 2024-04-28
"Material" is sometimes used as a synonym or generic term for (q3)shader. Maybe we could call these metamaterials?