@VReaperV 2024-04-28
> "Material" is sometimes used as a synonym or generic term for (q3)shader. Maybe we could call these metamaterials?

Yeah, I thought the name might cause some confusion, so a rename might be in order.