@illwieckz 2024-04-28
More details about what is probably the same of the first one of the previous comment:

```c
Unvanquished/daemon/src/engine/renderer/gl_shader.h: In constructor ‘GLUniform::GLUniform(GLShader*, const char*, const char*, GLuint, GLuint, bool, int, bool)’:
/home/illwieckz/dev/buildme-unvanquished/Unvanquished/daemon/src/engine/renderer/gl_shader.h:387:20: warning: ‘GLUniform::_global’ will be initialized after [-Wreorder]
  387 |         const bool _global = false; // This uniform won't go into materials SSBO if true
      |                    ^~~~~~~
Unvanquished/daemon/src/engine/renderer/gl_shader.h:381:19: warning:   ‘const int GLUniform::_components’ [-Wreorder]
  381 |         const int _components;
      |                   ^~~~~~~~~~~
Unvanquished/daemon/src/engine/renderer/gl_shader.h:392:9: warning:   when initialized here [-Wreorder]
  392 |         GLUniform( GLShader *shader, const char *name, const char* type, const GLuint std430Size, const GLuint std430Alignment,
      |         ^~~~
```