@illwieckz 2024-04-28
Also this:

```diff
diff --git a/src/engine/renderer/TextureManager.h b/src/engine/renderer/TextureManager.h
index d31892774..d28c78cb3 100644
--- a/src/engine/renderer/TextureManager.h
+++ b/src/engine/renderer/TextureManager.h
@@ -36,6 +36,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef TEXTURE_MANAGER_H
 #define TEXTURE_MANAGER_H
 
+#include <queue>
 #include <vector>
 #include <unordered_set>
 #include "GL/glew.h"
```

It is caught by older compilers (GCC9, ICC…):

```c
Unvanquished/daemon/src/engine/renderer/TextureManager.h:86:8: error: ‘queue’ in namespace ‘std’ does not name a template type
   86 |   std::queue<Texture*> texQueue;
      |        ^~~~~
Unvanquished/daemon/src/engine/renderer/TextureManager.h:42:1: note: ‘std::queue’ is defined in header ‘<queue>’; did you forget to ‘#include <queue>’?
   41 | #include "GL/glew.h"
  +++ |+#include <queue>
```

```c
Unvanquished/daemon/src/engine/renderer/TextureManager.h(86): error: qualified name is not allowed
  		std::queue<Texture*> texQueue;
  		^
Unvanquished/daemon/src/engine/renderer/TextureManager.h(86): error #77: this declaration has no storage class or type specifier
  		std::queue<Texture*> texQueue;
  		^
```