@VReaperV 2024-04-28
Yeah, I'll fix most of those later, when everything is more ready. I'm not worried about the ones that don't actually change how the code is executed right now.
The && within || warn is just bullshit however.