@VReaperV 2024-04-28
> With AMDGPU-PRO OGLP 23.40:
> 
> ```
> Warn: Compile log:
> ERROR: 0:94: 'gl_DrawIDARB' : required extension not requested: GL_ARB_shader_draw_parameters
> ERROR: 0:94: '' : missing #endif 
> ERROR: 0:94: '' : compilation terminated 
> ERROR: 3 compilation errors.  No code generated.
> Warn: Unhandled exception (15ShaderException): Couldn't compile vertex shader: generic2D 
> ```
> 
> https://pastebin.com/JXGTBaky
> 
> With Mesa radeonsi 23.2.1:
> 
> ```
> Warn: Compile log:
> 0:94(14): error: `gl_DrawIDARB' undeclared
> 0:95(20): error: `gl_BaseInstanceARB' undeclared
>  Warn: Unhandled exception (15ShaderException): Couldn't compile vertex shader: generic2D 
> ```
> 
> https://pastebin.com/0qNYBnX1

It looks like `AddExtension()` for `GL_ARB_shader_draw_parameters` will have to be set as version -1, because it relies on the variables/functions/etc in an extension having the same names as in the core version.