@illwieckz 2024-04-28
With both drivers case I get this though:

```
...using GL_ARB_shader_draw_parameters
```