@VReaperV 2024-04-28
> With both drivers case I get this though:
> 
> ```
> ...using GL_ARB_shader_draw_parameters
> ```

I'm pretty sure it works that way will all extensions that we use. But the shader never declares the extension as required in this case.