@illwieckz 2024-04-28
Anyway with `-1` I got this with Mesa radeonsi:

![unvanquished_2024-04-28_183358_000](https://github.com/DaemonEngine/Daemon/assets/2311649/70b7e425-d38f-45bf-9a2e-701371819fa7)

With OGLP the computer hangs for several minute, then later the game window blink forever with some textures painted in full screen, doing a screenshot just captures black.