@VReaperV 2024-04-28
Some graphics settings are currently broken on this branch, so might be that.