@illwieckz 2024-05-09
Graphics: `low` preset plus `r_normalMapping 1` and `cg_draw2D 0`.

This is framerate on my end using default spectator scene of the given maps.

map|master (fps)|material (fps)
-|-|-
atcshd|1500|1300
plat23|1600|1300
habitat|600|700
station12-b7|150|300

I noticed doing `r_showTris` on `master` branch that the station12 map has almost no vis optimization, it's like everything from the map is seen.